[feat]: interBackendAccess middleware
