+++ backend/loc/src/routes/query.ts
@@ -1,9 +1,10 @@
 require("dotenv").config();
 import express, { Request, Response } from "express";
-
+// import bcrypt from "bcryptjs";
 import z from "zod";
 import HttpStatusCode from "../types/HttpStatusCode";
 import { getPrivacyEntitiesWithVisibility } from "../query";
+import { interBackendAccess } from "../middlewares/interBackendAccess";
 
 /*
     {
@@ -33,77 +34,59 @@ const queryRouter = express.Router();
 
 queryRouter.use(express.json());
 
-queryRouter.post("/query", async (req: Request, res: Response) => {
-  const { authorization } = req.headers;
+queryRouter.post(
+  "/query",
+  interBackendAccess,
+  async (req: Request, res: Response) => {
+    console.log("The loc server is queried");
 
-  if (!authorization) {
-    return res
-      .status(HttpStatusCode.UNAUTHORIZED)
-      .send({ errors: "You are unauthorized to access this resource" });
-  }
+    const queryPayload = req.body;
+    const parsedPayload = queryRequest.safeParse(queryPayload);
 
-  if (!process.env.BACKEND_INTERCOMMUNICATION_SECRET) {
-    console.error(
-      "BACKEND_INTERCOMMUNICATION_SECRET is undefined. Check the .env",
+    if (!parsedPayload.success) {
+      res.status(HttpStatusCode.LENGTH_REQUIRED).json({
+        msg: "You sent the wrong inputs",
+      });
+      return;
+    }
+
+    const {
+      userId,
+      latitude,
+      longitude,
+      thresholdDistance,
+      college,
+      age,
+      gender,
+      isVisible,
+    } = parsedPayload.data;
+
+    const entityType = "user";
+
+    const privacyEntitiesWithVisibility =
+      await getPrivacyEntitiesWithVisibility(
+        latitude,
+        longitude,
+        thresholdDistance,
+        age,
+        college,
+        gender,
+        isVisible,
+        entityType,
+      );
+
+    if (!privacyEntitiesWithVisibility) {
+      res.status(HttpStatusCode.BAD_REQUEST).json({
+        msg: "The backend could not run queries on the location database",
+      });
+    }
+
+    res.status(HttpStatusCode.OK).json(privacyEntitiesWithVisibility);
+    console.log(
+      `PRIVACY ${gender} USERS NEARBY WITHIN ${thresholdDistance} meters and age less than ${age} with college ${college} and visibility set to ${isVisible}`,
     );
-    return res
-      .status(HttpStatusCode.INTERNAL_SERVER_ERROR)
-      .send({ errors: "Internal Server Error" });
-  }
-
-  const token = authorization.replace("Bearer ", "");
-
-  if (token !== process.env.BACKEND_INTERCOMMUNICATION_SECRET) {
-    return res
-      .status(HttpStatusCode.UNAUTHORIZED)
-      .send({ errors: "You are unauthorized to access this resource" });
-  }
-
-  const queryPayload = req.body;
-  const parsedPayload = queryRequest.safeParse(queryPayload);
-
-  if (!parsedPayload.success) {
-    res.status(HttpStatusCode.LENGTH_REQUIRED).json({
-      msg: "You sent the wrong inputs",
-    });
-    return;
-  }
-
-  const {
-    userId,
-    latitude,
-    longitude,
-    thresholdDistance,
-    college,
-    age,
-    gender,
-    isVisible,
-  } = parsedPayload.data;
-
-  const entityType = "user";
-
-  const privacyEntitiesWithVisibility = await getPrivacyEntitiesWithVisibility(
-    latitude,
-    longitude,
-    thresholdDistance,
-    age,
-    college,
-    gender,
-    isVisible,
-    entityType,
-  );
-
-  if (!privacyEntitiesWithVisibility) {
-    res.status(HttpStatusCode.BAD_REQUEST).json({
-      msg: "The backend could not run queries on the location database",
-    });
-  }
-
-  res.status(HttpStatusCode.OK).json(privacyEntitiesWithVisibility);
-  console.log(
-    `PRIVACY ${gender} USERS NEARBY WITHIN ${thresholdDistance} meters and age less than ${age} with college ${college} and visibility set to ${isVisible}`,
-  );
-  console.log(privacyEntitiesWithVisibility);
-});
+    console.log(privacyEntitiesWithVisibility);
+  },
+);
 
 export default queryRouter;
